Mod: `typeof_detail.js` module moved to `util/` - for organize modules
