+++ src/static/packet_conv/lanes_info.js
@@ -4,7 +4,7 @@
  * @author aKuad
  */
 
-import { typeof_detail } from "../typeof_detail.js";
+import { typeof_detail } from "../util/typeof_detail.js";
 import { LaneInfo } from "./LaneInfo.js";
