+++ src/static/util/typeof_detail.js
